modify the project structure
