+++ .travis.yml
@@ -6,7 +6,7 @@ install:
 - pip install -r requirements.txt
 
 script:
-  py.test --cov=api api/tests/
+  py.test --cov=api tests/
 
 after_success:
   - coveralls
