+++ models/users.py
@@ -1,0 +1,2 @@
+class User:
+    pass
