--- templates/401.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-     <head>
-        <meta charset="utf-8">
-        <title>Store Manager Version 1</title>
-        <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    </head>
-    <body>
-        <div class="container center-error">
-            <h1>Unauthorized access</h1>
-            <p>
-                You need to Login for you to access our resources.
-                </br>
-                Please proceed to <a href="{{ url_for('index') }}">Home Page</a>
-            </p>
-        </div>
-    </body>
-</html>
