--- templates/404.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-     <head>
-        <meta charset="utf-8">
-        <title>Store Manager Version 1</title>
-        <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    </head>
-    <body>
-        <div class="center-error">
-            <h1>Page Not Found</h1>
-            <p>The page you're looking for does not exist.
-            <p><a href="{{ url_for('index') }}">Return to home page</a>
-        </div>
-    </body>
-</html>
-
