--- templates/500.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-     <head>
-        <meta charset="utf-8">
-        <title>Store Manager Version 1</title>
-        <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    </head>
-    <body>
-        <div class="container center-error">
-            <h1>Server Error</h1>
-            <p>
-                An internal server error was encountered
-                </br>
-                Please try again later<a href="{{ url_for('index') }}">Home Page</a> 
-            </p>
-        </div>
-    </body>
-</html>
