+++ tests/test_api_get_all_sales.py
